commonutils/requirements.txt
setuptools==42.0.2
numpy==1.16.5
pandas==0.25.1
requests==2.22.0
Pillow==6.1.0
matplotlib==3.1.1
pymongo==3.9.0
boto3==1.9.200
redis==3.2.1
Flask==1.1.1
tqdm==4.32.1